+++ src/ladb_opencutlist/ruby/constants.rb
@@ -2,6 +2,6 @@ module Ladb::OpenCutList
 
     EXTENSION_NAME = 'OpenCutList'.freeze
     EXTENSION_VERSION = '1.5.1'.freeze
-    EXTENSION_BUILD = '201807172121'.freeze
+    EXTENSION_BUILD = '201807180626'.freeze
 
 end