+++ CHANGELOG.md
@@ -11,6 +11,7 @@ in 1.x and 0.x versions.
   * Hide cutlist group's raw dimensions if no length or width increase is defined
   * Added persistant material UUID mecanism to improve options storage
   * Added settings tab to manage dialog size and position
+  * Added cutting diagram options defaults storage
   
 * 1.5.0 (2018-06-25)